# swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/cars":{"get":{"operationId":"CarsController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Cars"]},"post":{"operationId":"CarsController_create","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"name":{"type":"string"},"plateNumber":{"type":"string"},"pricePerDay":{"type":"number"},"file":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":""}},"tags":["Cars"],"security":[{"bearer":[]}]}},"/cars/{id}":{"get":{"operationId":"CarsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Cars"]},"patch":{"operationId":"CarsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"name":{"type":"string"},"plateNumber":{"type":"string"},"pricePerDay":{"type":"number"},"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":""}},"tags":["Cars"],"security":[{"bearer":[]}]}},"/cars/{id}/deactivate":{"patch":{"operationId":"CarsController_deactivate","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Cars"],"security":[{"bearer":[]}]}},"/cars/{id}/activate":{"patch":{"operationId":"CarsController_activate","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Cars"],"security":[{"bearer":[]}]}},"/users":{"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]},"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Users"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/profile":{"get":{"operationId":"AuthController_getProfile","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/rentals":{"post":{"operationId":"RentalsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRentalDto"}}}},"responses":{"201":{"description":""}},"tags":["Rentals"],"security":[{"bearer":[]}]},"get":{"operationId":"RentalsController_findAllRentals","parameters":[],"responses":{"200":{"description":""}},"tags":["Rentals"],"security":[{"bearer":[]}]}},"/rentals/my":{"get":{"operationId":"RentalsController_findMyRentals","parameters":[],"responses":{"200":{"description":""}},"tags":["Rentals"],"security":[{"bearer":[]}]}},"/rentals/{id}/cancel":{"patch":{"operationId":"RentalsController_cancel","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Rentals"],"security":[{"bearer":[]}]}},"/rentals/{id}/complete":{"patch":{"operationId":"RentalsController_complete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Rentals"],"security":[{"bearer":[]}]}},"/payments/history":{"get":{"operationId":"PaymentsController_history","parameters":[],"responses":{"200":{"description":""}},"tags":["Payments"],"security":[{"bearer":[]}]}},"/payments/{id}/upload-proof":{"patch":{"operationId":"PaymentsController_uploadProof","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":""}},"tags":["Payments"],"security":[{"bearer":[]}]}},"/payments/{id}/confirm":{"patch":{"operationId":"PaymentsController_confirm","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Payments"],"security":[{"bearer":[]}]}},"/payments/{id}/reject":{"patch":{"operationId":"PaymentsController_reject","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Payments"],"security":[{"bearer":[]}]}},"/payments":{"get":{"operationId":"PaymentsController_findAllPayments","parameters":[],"responses":{"200":{"description":""}},"tags":["Payments"],"security":[{"bearer":[]}]}}},"info":{"title":"Car Rental API","description":"Documentation for Car Rental Backend API","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"name":{"type":"string","example":"Dope"},"email":{"type":"string","example":"dope@gmail.com"},"password":{"type":"string","example":"123456"}},"required":["name","email","password"]},"RegisterDto":{"type":"object","properties":{"name":{"type":"string","example":"Dope"},"email":{"type":"string","example":"dope@gmail.com"},"password":{"type":"string","example":"123456"}},"required":["name","email","password"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"dope@gmail.com"},"password":{"type":"string","example":"123456"}},"required":["email","password"]},"CreateRentalDto":{"type":"object","properties":{"carId":{"type":"number","example":1},"startDate":{"type":"string","example":"2026-06-10"},"endDate":{"type":"string","example":"2026-06-12"}},"required":["carId","startDate","endDate"]}}}}